feat: add reusable component preview
